requirements.txt
Django==5.2.5
requests==2.31.0
python-dotenv==1.0.0
whitenoise==6.5.0
gunicorn==21.2.0
discord.py==2.3.2
openai==1.3.0